+++ CHANGELOG.md
@@ -13,3 +13,6 @@
 - Inicialización de ``tiposIncidencia`` en `ServicioIncidencia` para añadir los tipos de incidencia por defecto en el sistema (Alberto Jiménez Expósito)
 - Añadida clase ``EstadoIncidencia`` que sirve para indicar los estados posibles de una `Incidencia` (Carlos Mayor Navarro)
 - Modificaciones en la declaración del atributo``estado`` dentro de `Incidencia` y métodos getter y setter, además del resto métodos relacionados con la misma en ``ServicioIncidencia`` (Carlos Mayor Navarro)
+- Modificación del método `nuevaIncidencia` de la clase ``ServicioIncidencia`` para implementar al completo la funcionalidad de incluir una nueva funcionalidad (Carlos Mayor Navarro)
+- Creación de los métodos `equals` y `hashCode` de la clase de la clase ``Incidencia`` para comparar dos incidencias por todos los atributos excepto del id (Carlos Mayor Navarro)
+- Creación de la clase ``IncidenciaYaRegistrada`` para lanzar una excepcion cuando ya exista una ``Incidencia`` ya registrada (Carlos Mayor Navarro)